Thm.sml changed to fix "le bug de Barras":
   load "bossLib";
   open bossLib;

   val F_DEF = Define ` f x y = T`;

   val th1 = PROVE [F_DEF] ` f (\x:num.0) (\x:num.1) `;
   val th2 = DECIDE ` 0=0 `;

   val th = MK_COMB(th1,th2);
   val abs = REWRITE_RULE [] th;

This is the first soundness bug in hol90/hol98 since 1990 or so (I think
the last one was found by Kevin Mitchell of Edinburgh; it had to do with
renaming). The actual bug is in MK_COMB -- a "derived" rule -- and it
occurs because the first argument to MK_COMB is not being checked to see
that it is an equality. All in the name of speed, of course.

Kudos to Bruno Barras, who found the bug and a nice demonstration of it.