Calls to Term now trap the PrecConflict error that can be raised by the parser as it constructs its precedence matrix.  The messages are better but still not really much help.
Also, fixed the pretty-printer to be more intelligent about quoting things
with $s.  In the old system (both in the Taupo development and before it)
``f $=> /\ p`` could be parsed, but would be printed as ``f => /\ p``,
which would ensure that what was printed couldn't be parsed back.
That behaviour is now fixed.