Fairly extensive changes, implementing a scheme whereby each theory keeps track of its own personal grammar value, which can then be relied on by other libraries.  This is in contrast with the term_grammar() value.  This gets changed all the time, and library implementations can't really rely on it being sensible.
For example, numLib quotes a term involving >.  If this symbol was over-
ridden or messed with in some other way, before numLib was loaded, it
would try to parse that term and come a cropper.  No more.

Changes in Preterm are more to do with trying to improve the error
reporting from the type-checking, particularly as with regards numerals.