1. Added a termination interface to Defn. This takes a defn and attaches a new goalstack to the list of current proofs.  There is also an analogue that wraps up a defn and a tactic and applies the tactic to the termination constraints of the definition. If the tactic proves the goal, then the unconstrained recursion equations and induction theorem are returned.
    tgoal  : defn -> proofs
    tprove : defn * tactic -> thm * thm

In tprove, some cleverness is used to arrange that the ultimate goal
(namely the unconstrained recursion equations and induction theorem) is
not what is initially seen by the user; instead the first goal appears
to be something of the form

     ?R. TC_1 /\ ... /\ TC_n /\ WF R

Proving this will deliver the top_thm()

      |- rec. eqns  /\  ind. thm

One problem with this sleight of hand is its behaviour with respect to
"restart" -- restarting takes us to the ultimate goal. To regain the
original state, Defn.TC_TAC <defn> may be used.

2. Halts.TC_SIMP_CONV uses a standard set of theorems about WF to simplify
goals ensuing from giving the witness for R above. Typically, one would
work by going

   val foo_defn = Hol_fun "foo" <eqns for foo> ;

   Defn.tgoal foo_defn;
   e (EXISTS_TAC <some termination relation>);
   e (CONV_TAC Halts.TC_SIMP_CONV);