Ho_theorems has long been a nasty abomination; its theorems should be stored in a theory file, not reproved each time.  Though theory specific grammars protect us from some dangers, they can't prevent the following scenario:
- load "arithmeticTheory";
> val it = () : unit
- new_definition ("f", Term`f x = x + 1`);
> val it = |- !x. f x = x + 1 : Thm.thm
- load "simpLib";

Type inference failure: the term

f

can not be constrained to be of type

:'a -> 'b

unification failure message: unify failed
! Uncaught exception:
! HOL_ERR <poly>
! Uncaught exception:
! Fail "load: unable to load"

This is caused because the script inside Ho_theorems uses "variables"
called "f" and "g" all over the place, but if these have to be parsed
as constants, the script will fail.  An ideal solution would be to be
able to do something analogous to the way grammars are handled: to have
constant name resolution take a parameter representing the names
that are to be resolved into constants, but that would require more
work and I've long wanted to turn this into a theory anyway.

Eventually, Ho_bool should almost certainly be merged into bool.
Ho_theorems now has a rather simpler implementation; it exports the
same values though, and also causes the same side effects (to Ho_rewrite's
implicit rewrites) when loaded.