Improve typesetting.  Now recognises every symbol.
Add support for HOLDoc parser directives of form

  (*[ NAME arg arg arg ]*)

(which appear as comments to HOL).

At the moment, implement just one: in a list of binders, eg. between
`!'  and `.', can write (*[ VARS a b c ]*) to add to the list of
identifiers that HOLDoc will count as variables in the current rule.
This is useful for set comprehensions, like in getifaddrs_1, where
HOLDoc has no idea of which variables are bound.

I have added this directive to getifaddrs_1.