Add IP constructor to typesetting; add orderings auxiliary to typesetting; add doq.
Deal better with FAIL, MEM, APPEND.

Fix TeXing of geterr comment.

Add i1, i2 to vars of delivery_out_1.

When errors are generated by the munger, don't do sort -u, do sort |
uniq -c so we can see how many repetitions of each error there are.

Add flag to HOL_CURRIED_ALIST that specifies whether we want parens to
be stripped and commas turned to }{ or not.

Fix thinko: an id that is an arg of a function may be terminated with
a separator, not just with white space.

Put the warning into the autogenerated file as well as stderr.

Add general two-arg spacing-fixing macro \Mspii{x}{y} to ltsmunge, for
use in TNet.imn.