Implementation of "parse only" type abbreviations.  The main function to use is type_abbrev : (string * hol_type) -> unit
If the hol_type has type variables in it, then these become parameters to
the string, making it an abbreviation for a type operator.  E.g.,
  type_abbrev("set", ``:'a -> bool``)
allows you to write ``:num set``.

No attempt to pretty-print types in accordance with stored abbreviations,
so the system will spit back ``:num -> bool`` in response to this.  (Hey,
if it's good enough for Moscow ML, it's good enough for us.)

Also split off type_grammars into their own source module.