Change to the start-up behaviour of HOL. At configuration time the following 4 scripts get placed in the bin directory (as well as a few others, e.g., Holmake):
     hol        -- loads and opens bossLib
     hol.bare   -- loads and opens boolLib only
     hol.unquote     -- the hol script, input passed through quotation filter
     hol.bare.unquote -- the hol.bare script, using quotation filter

Most of the time, the hol or hol.unquote scripts offer the "right" context
for people to start working in.

This is tentative: I may change the names of the scripts
(suggestions welcome), 	or increase the libraries that the
hol/hol.unquote scripts load and open by default.

config-muddy.sml has been incorporated into configure.sml