Introduction of strip_binder into kernel. This allows many consecutive binders to be stripped off efficiently. It takes about 18 seconds to build an abstraction with a million binders, and about 26 seconds to take it back apart.
Users may notice that the prettyprinter is lying to them. This is
because the prettyprinter currently iterates "dest_abs" to break terms
down, which can give different renaming behaviour that the bulk
processing of strip_{abs,forall,exists}. It was too hard to fully
simulate the processing of iterated dest_abs.