Following a suggestion of mjcg, I have made Define and xDefine fail when they encounter a free variable that occurs on the right hand side, and not the left.
For defining schematic definitions, the flag Globals.allow_schema_definition
needs to be set. It is initially false.

There are also entrypoints in TotalDefn, DefineSchema and xDefineSchema,
which take care of setting and unsetting the flag when making a
schematic definition.

This is not yet fully worked out for the schematic definitions, but
ordinary ones should work as expected.