Modifications to DB. It now offers separate calls to get the theorems, definitions, and axioms from a theory.
I also realized that "theorem" wasn't precise. This is because the
database conflates theory elements whose names are equal when
translated to lower case. This is useful for DB.find.  Hmm, must
think about it ... offhand, I think I'll just make a note to that
effect in the documentation. If people want real precision, they'll
have to use ML identifiers. Still, the database is for helping people
find stuff: if something's not there, chances are slimmer that they'll
find it!