Revisions to type and term matching. Part code clean up and maybe some bugs have been fixed too.  The changes were prompted by Mike G. finding some differences with some old code.
There is a difference with old code, namely that the matching algorithms
take sets of type variables and variables that are not allowed to be
instantiated in the course of matching. This serves as a cut-off for
some proof procedures. A scenario would be one where a match succeeded,
but subsequent instantiation failed because variables moved in the
resulting substitution were actually free on the assumption list of the
theorem being instantiated.

However, used by themselves, with non-empty "avoid" sets, one could spot
a difference between old and new versions of matching. I'd be interested
to see such examples.