My "solution" to dealing with the functor, which is morally the right way to do this.  To get the dependencies to work properly for theory construction, you have to write a rule for xTheory.sml, because the logic in Holmake treats the intermediate stages (xScript.uo and the executable) as hidden parts of the atomic step from Script.sml to Theory.sml.
This solution is not ideal for a couple of reasons:
  * stray word32Script.exe and word32Script.{ui,uo} files are left hanging
    around.
  * you have to write a relatively complicated rule to get the construction
    of the theory to happen.
Though the previous approach spat out those warning messages about
wordFunctor being used in the wrong way, it was simpler.

I've also changed the parameter passed to the functor to be the number of
bits in the words; this seems simpler.