Added support for boolification of user-defined datatypes.
This is done by some "polytypic programming" in Datatype.sml copying
Konrad's existing code for generating size functions.  However, since
it makes heavy use of lists, and lists come after Datatype in the
build sequence, this probably isn't the right place for the code to
live.

After discussing this with Michael, it would probably be best to move
the code to a module in src/Boolify/src, and register it using the
hook mechanism in the TypeBase.

Note that src/Boolify/src comes immediately after tfl in the build
sequence, because it defines some functions (such as num_to_bool)
using non-primitive recursion. Even if this could be "done by hand",
decoding boolean lists into datatypes will undoubtedly involve defining
some functions with hairy termination properties.