Changed the type (and nature) of register_update_fn in TypeBase.
Before, you registered a function that got to see a tyinfo before it
was added to the TypeBase. All well and good.

Now, you register a function that gets to modify a list of tyinfos
before they're added to the TypeBase. Distinctly hairier.

Why allow changes? Well, how else can I register a piece of code to
add "boolify" entries to the tyinfos.

Why a list of tyinfos, instead of one at a time? Well, with mutually
recursive datatypes, you really need to see all of the tyinfos at the
same time.

Note that now the order of execution of these update functions is
significant. They get executed in the order that they're registered,
so something registered after boolification gets to see the "boolify"
entries and make use of them.