This fixes a pretty obscure bug:
   g `(h::t = x::y) /\ (SUC z = SUC w) ==> P x y`;
   e (BasicProvers.PRIM_STP_TAC boolSimps.bool_ss NO_TAC);

goes into an infinite loop.  The reason is that after simplification, the
goal looks like

    P x y
  ----------
    0. h::t = x::y
    1. SUC z = SUC w

The code for PRIM_STP_TAC then sucks those assumptions that are equations
between constructors off the list and puts them back on after simplifying
with the simpset (bool_ss in this case).  Because bool_ss doesn't affect
either, they are put back unchanged, but in the wrong order.  The controlling
logic in PRIM_STP_TAC repeats this process under a CHANGED_TAC.

Arguably the fault lies in the logic that assumes the given simpset will
always be able to modify (for the better) equations between constructors, but
fixing this seems harder.  Instead, my fix just makes sure that the relevant
assumptions get put back on in the right order.