Added wordFunctorLib.sml.
This enables the word library file (for word conversion and
pretty-printing) to be created in the same way as the theory file.

A program mkword.exe can now be used to produce n-bit words.
For example, mkword.exe 4, produces 4-bit words.  A -sys option
copies the object files to the SIGOBJ directory.

The word32/ directory is to be deleted, being replaced with n_bit/.

Anthony