Fixed infelicities in term parser location management; all ranges should come out correctly now.
I've also tidied up the code for applying a rule rather a lot, and I
believe fixed a latent bug: when parsing a listfix operator, if a TM
immediately precedes the left delimiter, it will be included in the
list!  This doesn't normally happen in HOL, since we write ``foo with
<| ... |>`` rather than ``foo <| ... |>``, but it could have surfaced
one day.