Allow "..", "...", "...." etc as symbols in HOL munger.
Fix Makefile to not specify a version of ocaml.
Fix Makefile to work with ocamldep v3.06 (weirdly, it no longer handles .mll files).

Attempt (but fail) to fix the code that matches ordinary TeX source,
to allow ([[x]]) to parse correctly.  Failed to do it because
complementing regexps is too hard for me to do on paper - anyone else
is welcome to try if they want!