Fix to remove an exponential blow-up when traversing a term with congruence rules.  The problem is that when traversal with congruence rule 1 fails to change the term, traversal with the next rule is attempted. But, when looking at a sub-term within the second traversal, both congruence procedures will be tried once more.   You can see the badness with
  - set_trace "simplifier" 2;
  - SIMP_CONV bool_ss [] ``p ==> q ==> r``

You will see the line

  more context:  [.] |- q

twice.  This problem is hard to notice in normal contexts because the only
built-in congruence rules are for implication and conditionals, and these
are not usually nested very deep.  However, big conjunctions and a rule for
conjunction combine to produce very bad performance.  Till now anyway!

Now, congruence procedures that "work", but which do not change the term,
mask other congruence procedures.