A backwards compatible fix for a curious problem that arises with the name-carrying implementation.  The problem is caused by the fact that if there's a term which prints as
  P l /\ ?l'. Q l l'

but where the dB implementation actually wants the l' to be l, then
after a case split on the free variable l (causing it to go away and
be replaced by h and t), the dB implementation will be able to "change
the l' back" to l.  Thus Joe's proof can do successive case splits on
l.  In the name-carrying implementation, there's no way for the bound
and free variables to have the same name.