An efficiency improvement for enumerated types.  Here's a little table giving times taken (in seconds) to define an enumerated type of 50 elements:
             dB indices      Name-carrying terms
old code      5.5               15.1
new code      4.0                3.9

I don't know why the old code and name-carrying terms interacted so badly,
but will look into this as it may point to a problem in the experimental
kernel.

For what it's worth, the performance graphs for the three good performances
are all linear-looking (for sz = 1..50), but the gradient on the line
for old code is slightly worse than that for the new code.