I should have mentioned LET_ss in the previous message. It implements applicative order simplification for let expressions. Unfortunately, it can't use the conceptually simple congruence rule   (v = v') ==> (LET f v = f v') because of a strange interaction between HO_PART_MATCH and the exact order that the simplifier takes in traversing terms.  Instead, a rather more involved two stage process is used:   (v = v') ==> (LET f v = LET f (I x)) is used as a congruence, evaluating v to v', and recording that this has been done by wrapping in the I combinator.  (You could use a special purpose marker term for this if you really wanted.)  Then, there's a rewrite    LET f (I v) = f v which will subsequently cause a beta-reduction (assuming that f is an abstraction).
The change to pairSimps is to replace its let-simplification with the
above technology.  What was normal-order becomes applicative order
with all the usual issues.  I think the gain in efficiency in most
cases is probably worth it.  I also wonder why bool_ss doesn't have
any treatment of LET, but I'll leave that stone unturned for now.