A while back, I made the simplifier treat a theorem [.] |- F as the rewrite [.] |- x = F.  This explains the difference between
 time
  (SIMP_CONV (pureSimps.pure_ss ++ numSimps.REDUCE_ss ++ boolSimps.BOOL_ss) [])
  ``~T ==> P (2 EXP 40)``;

and

 time
  (SIMP_CONV (pureSimps.pure_ss ++ numSimps.REDUCE_ss ++
              boolSimps.BOOL_ss ++ boolSimps.CONG_ss) [])
  ``~T ==> P (2 EXP 40)``;

In the latter, ~T is rewritten to false, so the congruence rule (in
CONG_ss) for implications lets false be assumed when the consequent is
rewritten.  This then lets P (2 EXP 40) to be rewritten to F without
having to calculate 2 ^ 40.

However, if given a rewrite of the form |- P ==> F, the old code would
rewrite this to |- P ==> (x = F), and because this applies to every
possible boolean term, a lot of time was spent trying to discharge the
side condition P.  So, the fix is to spot P ==> F, for any P, and to
rewrite this to ~P.  (This will later get rewritten to either P = F,
or if P is an equality to both P = F, and also (SYM P) = F.)