Import an example of the use of HOLDoc for typesetting a labelled transition system.
Net_Hol_reln.sml,.sig was written by Michael Norrish for the NetSem
project.

ltsmunge.sty was written by Keith Wansbrough and Peter Sewell for the
NetSem project.

The contents of this directory are intended to serve as an example and
as a template for others to use HOLDoc.  Feel free to use them in
whole or in part.