A fix for a bug in the simplifier's treatment of congruence rules that prevented it from handling "higher order" rules, like this one for RES_FORALL:
  |- (P = Q) ==> (!x. x IN P ==> (f x = g x)) ==>
     (RES_FORALL P f = RES_FORALL Q g)

Feeling brave, I then put this, and the accompanying rule for RES_EXISTS
into bool_ss.  Congruence rules take precedence over normal term traversal
in the simplifier, so with a rule like the above installed, the constant
RES_FORALL is never "seen" on its own.  In particular, this means that
you can't rewrite with RES_FORALL_DEF, which looks like

  |- RES_FORALL = (\p m. !x. x IN p ==> m x)

Instead, you need to rewrite with something that matches at the same level
(or higher) as the congruence rule.  Thus, RES_{FORALL,EXISTS,SELECT}_THM
in boolTheory.  These can then be the basis for the theorems in
res_quanScript, and look like

  |- RES_FORALL P f = !x. x IN P ==> f x