Added ability to export ML files. Most of the desired functionality is now present. Still to do: (1) figure out a scheme for mapping from HOL's "unrestricted" polymorphism to ML's restrictions, i.e., eqtypes. For example, consider the definition of "mem"
   mem x [] = T
   mem x (h::t) = (x=h) \/ mem x t

In HOL, mem has type :'a -> 'a list -> bool. In ML, the type is
:''a -> ''a list -> bool. For functions defined with equality tests in
the rhs, I can see how to find out what the resulting ML type should be
(by examining the types of the occurrences of equality). But the stumper
is when there's a layer of indirection. E.g,

    insert x l = if mem x l then l else x::l

should get ML type :''a -> ''a list -> ''a list, but it seems I will need
to maintain some kind of information about defined functions.

The other thing still to do is to teach Holmake about the exported ML
files. Roughly, from an xScript file, there should (always) be generated
an xTheory.{sig,sml} pair, and (possibly) an xML.{sig,sml} pair. ("Possibly"
because a theory may not in fact have any executable definitions.) This
may be easy, but it would be nice if it was sufficiently general that other
kinds of files automatically generated from HOL scripts could also
have their dependencies maintained.