Possibly a fix to an obscure MacOS bug, but also possible that it's simply too late and I'm hallucinating. I was getting different behaviour when invoking bin/build twice in a row. This was traced to the removal of all symbolic links from the sigobj directory. Sometimes they would all get removed, and the next time only about half of them would be removed.
 The change, which fixes it, and I don't know why, is to change
the "map_dir" function which maps a function over all files in a
directory. Specifically, we now open the directory stream and
accumulate all file names into a list. Then the (argument) function
is mapped over the list. Previously, the function was applied to
each file as it was pulled off the directory stream. But it seemed
as though the directory stream was affected by the application
of the function (which was removing the named files). Could the
deletion of a large number of files from a directory cause the
remaining files in the directory stream to disappear?

Hrm.