HOLDoc:
Field names are often used as variable names, so the class FIELD is treated specially.

This handling doesn't work very well for field names containing
underscores, as it interacts with the subscript handling for variable
names.

This commit makes it marginally better in certain cases.  In
particular, you should never see a \tsfield{} in the output.