HOLDoc:
When attempting to render a variable name in subscripted form, say
x_y, don't assume that if y is a field name it is a variable name - it
must be explicitly declared as a variable name.  x however may be
either a field name or a variable name.

This means cb_tf_acknow, as used in the TCP spec, renders as a single
variable, rather than as cb_{tf_acknow}, which is bogus.