HOLDoc:
Allow a variable or field named "is2" to be typeset as is_2 when
SMART_PREFIX is on.  The earlier fix didn't work, because the case
here eliminated was matched first.  Fortuitously, the new logic also
prevents the x_y case that this case was intended to fix.  This is
however likely to break any use of subscripts with underscores.  Let
me know if there are problems.