Allow differently-based numeric literals.  Can now write   0xAF or   0b1010011 or   063 The latter is octal and requires the flag in base_tokens to be set to true.  (This is a backwards compatibility bodge, just in case there are people up there who liked leading decimal zeroes.)
Also, all numeric constants can be intermingled with underscores, which
are ignored.  These can be used to space numbers for extra legibility.
For example, 12_345_678 is easier to spot as 12 million than 12345678.

Note the convention that numeric type identifier characters don't overlap
with hex digits and are lowercase.  For example, 0xAAw would be the word
170.  In the potentially ambiguous case of 0xaa, the a is taken to be a
hex digit, not a numeric type specifier.