Booth multiplier example. Original code by Anthony Fox (part of his ARM6 verification) was modified by Juliano Iyoda for hardware synthesis.
The example is not yet complete but is being used by MJCG to
stress-test the hardware compiler.

There is a weird bug causing compilation (Holmake -I word32 -I .. -I
../dff) with the following error.

Exporting theory "boothDev" ... done.
Theory "boothDev" took 62.022s to build
Analysing boothDevTheory.sml
Analysing boothDevTheory.sig
File "boothDevTheory", line 493, characters 78-79:
!                                                                  count1).
!                                                                         ^
! Lexical error: ill-formed token

Holdep failed.


We are not sure if this is a problem in the example, or an obscure HOL
bug.