----------------------------------------------------------------------
Edited the "real" library to use the new quotient library.

The hrealScript.sml and realScript.sml files only had a mention of
EquivType in a comment, not for actual use; these have been deleted.

The hratScript.sml and realaxScript.sml actually used the EquivType
tool to do quotients; this has now been revised to use the quotient
tool.  In each case one new well-formedness (respectfulness) theorem
had to be proven and included as an argument to the tool.

In seqScript.sml, there was a small incompatibility where a definition
of a new constant named "-->" conflicted with the constant "--> defined
in the quotient library.  This was healed by hiding the previous definition
before making the new one.

All of this was very straightforward.

Modified Files:
 	hol98/src/real/hratScript.sml hol98/src/real/hrealScript.sml
 	hol98/src/real/realScript.sml hol98/src/real/realaxScript.sml
 	hol98/src/real/seqScript.sml
----------------------------------------------------------------------