----------------------------------------------------------------------
 Committing in the wordFunctor.sml file of the n_bit library.

This has now been modified to use the new higher order quotients
library instead of EquivType.

The source code involved a good deal of messy probing into the
internals of the quotient operation, which was dependent on the
previous implementation.  This has been clarified, but the current
file is now dependent somewhat on the implementation of the new
quotient library, and so would not be expected to build if the
references to open "quotient" were changed to open "EquivType".

There were several places where the source code manually dealt
with the Hilbert choice operator, also referred to as the "SELECT"
operator.  These have been elided, since the new quotient package
provides an encapsulation of these details.  Many proofs have been
simplified or eliminated.  In particular, the tactic "SELECT_WORD_TAC"
has been changed to the simpler "QUOTIENT_WORD_TAC", which is used
repeatedly.

This completes the set of modifications to parts of the standard
HOL system that used EquivType.  I will also examine the examples
directory to see if there are any dependencies there.

 Modified Files:
 	hol98/src/n_bit/wordFunctor.sml
 ----------------------------------------------------------------------