Added BddAssume and BddDisch, the ASSUME and DISCH analogs for term_bdd.
This means we can do more with the assumptions of a term_bdd than just watch
them get passed around. Specifically, now I can discharge assumptions to a
term_bdd (needed this for unlazification). BddAssume has been added for completeness.

BddDisch had to be a primitive, since it needs to delete from the assumption set
and there was no derived way to do that. BddAssume is currently implemented as a derived rule,
though efficiency considerations may change that.