---------------------------------------------------------------------- Enter Log.  Lines beginning with `CVS:' are removed automatically
Committing in a revision to the quotients library, with associated
corrections to the dependent libraries.

Added a feature requested by Michael Norrish, that the tool generate
trivial respectfulness theorems if the user leaves them out, of the
form
    !x y. (x = y) ==> (treal_of_hreal x) treal_eq (treal_of_hreal y)

where the relation is a true equivalence relation, i.e., reflexive.

This enabled the deletion of three trivial respectfulness theorems
that I had added when moving the dependent libraries to use the new
quotient tool.  Harrison's tool had not required these, because his
system only dealt with true equivalence relations, which are always
reflexive, so in the above example, the respectfulness of treal_of_hreal
was easy to prove.

Also added a feature to check that all of the constants being lifted
are properly represented by respectfulness theorems.  This is only
sensible, and should catch a simple and common user error.

Also divided the quotientScript.sml file into five, adding four
new files for the four type operators "list", "option", "pair", and
"sum".  Each has the definitions and theorems appropriate for that
type operator, and for some of the common constants associated with
the type operator.

Also added quotientLib.sml, which includes all the others.  Thus a
user can easily get everything in the quotient library by simply
typing "open quotientLib;".

Modified Files:
	hol98/src/integer/integerScript.sml
	hol98/src/n_bit/wordFunctor.sml
	hol98/src/quotient/src/quotient.sig
	hol98/src/quotient/src/quotient.sml
	hol98/src/quotient/src/quotientScript.sml
	hol98/src/real/hratScript.sml hol98/src/real/realaxScript.sml
Added Files:
	hol98/src/quotient/src/quotientLib.sml
	hol98/src/quotient/src/quotient_listScript.sml
	hol98/src/quotient/src/quotient_optionScript.sml
	hol98/src/quotient/src/quotient_pairScript.sml
	hol98/src/quotient/src/quotient_sumScript.sml
----------------------------------------------------------------------