----------------------------------------------------------------------
Committing in definitions.tex, with an added section that describes
briefly the quotient package.  Much of this is a repeat of section 7
of the paper, ``Higher Order Quotients in Higher Order Logic,''
where almost every individual argument is described more fully in
some section of the paper.  However, it would have been disproportionate
to copy more of the paper in here.  This is enough to give someone a
taste for what the quotients package can do, or to give someone who had
used it before a reminder as to the exact form of each of the arguments.
For more information, or for more extensive use, they really should
look at the paper.

By the way, I found that the entire Description document does not build
with the given Makefile.  I needed to change the line

description.dvi: description.tex $(CHAPTERS) ../LaTeX/commands.tex description.idx
to
description.dvi: description.tex $(CHAPTERS) ../LaTeX/commands.tex

in order for the Description document to build from the bare download.
May I suggest that either Michael or Konrad take a look at fixing this.

Modified Files:
	hol98/Manual/Description/definitions.tex
----------------------------------------------------------------------