----------------------------------------------------------------------
Committing in the source of documentation for the quotients library.
Added the .tex source, with the necessary llncs.cls file, and also
made a Makefile for easy generation of the document.  Left the .ps
and .pdf documents available (updated), in case someone does not
get around to making the src/quotient/doc directory.

Modified Files:
	hol98/src/quotient/doc/quotient2.pdf
	hol98/src/quotient/doc/quotient2.ps
Added Files:
	hol98/src/quotient/doc/Makefile
	hol98/src/quotient/doc/llncs.cls
	hol98/src/quotient/doc/quotient2.tex
----------------------------------------------------------------------