Follow-on to the change to the pre-processor. There's a significant problem, in that the heretofore standard way of protecting a piece of code from alterations to the grammar(s) used to parse terms in that piece of code ... no longer works after the change.
The previous recommendation was to re-bind Type and Term with

  val (Type,Term) = Parse.parse_from_grammars <grms>

which allowed subsequent

   Term ` .... `

to parse with the specified grammar. However, also

   ``  ... ``

would subsequently use the specified grammar as well. But I
changed the pre-processor, and this broke the usage of

   `` ... ``

with the specified grammar.

  The fix is to do as in src/real/RealArith.sml: get the
ambient grammar, set the desired grammar, execute the
body of code, then reset to the ambient grammar. Works
provided the code doesn't alter the grammar.

  This probably needs to be documented in a more visible place,
once Michael comments.