----------------------------------------------------------------------
Committing in a revised version of the quotients package, and an example.

The quotient theorems are broken into theories based on the type operator
they refer to, with general non-specific theorems remaining in
quotientScript.sml.

The example of creating finite sets from lists by a quotient is presented
in the new "examples" directory, in finite_setScript.sml.  It is perhaps
surprisingly simple to walk through.

Modified Files:
	hol98/src/quotient/src/quotient.sml
	hol98/src/quotient/src/quotientLib.sml
	hol98/src/quotient/src/quotientScript.sml
	hol98/src/quotient/src/quotient_listScript.sml
	hol98/src/quotient/src/quotient_optionScript.sml
	hol98/src/quotient/src/quotient_pairScript.sml
	hol98/src/quotient/src/quotient_sumScript.sml
Added Files:
	hol98/src/quotient/examples/finite_setScript.sml
	hol98/src/quotient/examples/ind_rel.sig
	hol98/src/quotient/examples/ind_rel.sml
	hol98/src/quotient/src/quotient_pred_setScript.sml
----------------------------------------------------------------------