----------------------------------------------------------------------
Committing in two new examples of using the quotient package.

The first is a very simple example, duplicating Paulson' most complex
example from his quotients paper.

The other is the most complex example so far, creating the syntax of
Abadi and Cardelli's sigma calculus, and proving the Church-Rosser
theorem for this language.  The quotient package is used in the file
"liftScript.sml".  This is the example from the 45-page paper in the
"doc" subdirectory.

To build either of these go to the directory where the example is
located and type Holmake.

Modified Files:
	hol98/src/quotient/src/quotient.sml
Added Files:
	hol98/src/quotient/examples/msgScript.sml
	hol98/src/quotient/examples/obj/MutualIndThen.sig
	hol98/src/quotient/examples/obj/MutualIndThen.sml
	hol98/src/quotient/examples/obj/alphaScript.sml
	hol98/src/quotient/examples/obj/barendregt.sig
	hol98/src/quotient/examples/obj/barendregt.sml
	hol98/src/quotient/examples/obj/ind_rel.sig
	hol98/src/quotient/examples/obj/ind_rel.sml
	hol98/src/quotient/examples/obj/liftScript.sml
	hol98/src/quotient/examples/obj/more_listScript.sml
	hol98/src/quotient/examples/obj/more_setScript.sml
	hol98/src/quotient/examples/obj/objectScript.sml
	hol98/src/quotient/examples/obj/reductionScript.sml
	hol98/src/quotient/examples/obj/semanticsScript.sml
	hol98/src/quotient/examples/obj/tactics.sig
	hol98/src/quotient/examples/obj/tactics.sml
	hol98/src/quotient/examples/obj/variableScript.sml
----------------------------------------------------------------------