---------------------------------------------------------------------- Enter Log.  Lines beginning with `CVS:' are removed automatically
Committing in a slight modification to the Makefile for the n_bit
library.  Apologies to Anthony Fox, but I found that the result of
running the existing Makefile was a PDF file with very chunky looking
characters.  I added an "if" condition to try using pdflatex first,
but if that failed, to use dvipdf as before.

Anthony, if you don't like this, please go ahead and change it back.
I'm just trying to help.

Modified Files:
	Manual/Makefile
----------------------------------------------------------------------